Exit from test script early if transpile fails

+++ scripts/test.sh
@@ -1,5 +1,9 @@
 #!/usr/bin/env bash
 npm run build
+build_response=$?
+if [[ ${build_response} -ne 0 ]]; then
+  exit ${build_response}
+fi
 
 rm -rf ./test/test_output
 mkdir ./test/test_output
@@ -7,8 +11,8 @@ mkdir ./test/test_output
 ./bin/tarr -d ./test/test_output -r ./test/routes.json
 
 npm run test
-ex=$?
+test_response=$?
 
 rm -rf ./test/test_output
 
-exit ${ex}
+exit ${test_response}
